+++ bootique-cayenne-jcache/src/main/java/io/bootique/cayenne/jcache/CayenneJCacheModuleProvider.java
@@ -1,0 +1,12 @@
+package io.bootique.cayenne.jcache;
+
+import com.google.inject.Module;
+import io.bootique.BQModuleProvider;
+
+public class CayenneJCacheModuleProvider implements BQModuleProvider {
+
+    @Override
+    public Module module() {
+        return new CayenneJCacheModule();
+    }
+}