+++ bootique-cayenne-jcache/src/test/java/io/bootique/cayenne/jcache/persistent/Table1.java
@@ -1,0 +1,9 @@
+package io.bootique.cayenne.jcache.persistent;
+
+import io.bootique.cayenne.jcache.persistent.auto._Table1;
+
+public class Table1 extends _Table1 {
+
+    private static final long serialVersionUID = 1L; 
+
+}