+++ RELEASE-NOTES.md
@@ -7,8 +7,8 @@
 * #27 Unit tests: add a builder of Table objects based on DbEntity
 * #28 Extend TestDataManager for Cayenne
 * #30 Add a way to register a listener of schema creation event
-* #31 New module: bootique-cayenne-ehcache
-* #32 bootique-cayenne-ehcache: enable cache refresh on commit
+* #31 New module: bootique-cayenne-jcache
+* #32 bootique-cayenne-jcache: enable cache refresh on commit
 * #33 Remove 'cayenne.config' configuration deprecated in 0.14 
 
 ## 0.17